フレンド削除の際にアラート表示

+++ client/Hi/Hi/View/Home/HomeView.swift
@@ -13,7 +13,9 @@ struct HomeView: View {
     @State private var searchText = ""
     @FocusState private var isFocused: Bool
     @ObservedObject var viewModel2: DeleteViewModel
-
+    @State private var isAlertPresented = false
+    @State private var userToDelete: User?
+    
     var filteredFollowings: [User] {
         if searchText.isEmpty {
             return userEnvironmentData.user.followings
@@ -21,54 +23,54 @@ struct HomeView: View {
             return userEnvironmentData.user.followings.filter { $0.name.localizedCaseInsensitiveContains(searchText) }
         }
     }
-
+    
     var body: some View {
         VStack {
             HStack {
                 Image(systemName: "magnifyingglass")
                     .foregroundColor(.gray)
-                    TextField("検索", text: $searchText)
-                        .focused($isFocused)
-                        .toolbar {
-                            ToolbarItemGroup(placement: .keyboard) {
-                                Spacer()
-                                Button("閉じる") {
-                                    isFocused = false
-                                }
+                TextField("検索", text: $searchText)
+                    .focused($isFocused)
+                    .toolbar {
+                        ToolbarItemGroup(placement: .keyboard) {
+                            Spacer()
+                            Button("閉じる") {
+                                isFocused = false
                             }
-                        }
-                        .onSubmit {
-                            if searchText.isEmpty {
-                                print("入力なし")
-                            } else {
-                                print("\(searchText)")
-                                let isFollowing = userEnvironmentData.user.followings.contains { user in
-                                    user.name == searchText
-                                }
-                                print(isFollowing)
-                            }
-                        }
-                    if isFocused {
-                        Button(action: {
-                            self.searchText = ""
-                        }) {
-                            Image(systemName: "multiply.circle.fill")
-                                .foregroundColor(.gray)
                         }
                     }
+                    .onSubmit {
+                        if searchText.isEmpty {
+                            print("入力なし")
+                        } else {
+                            print("\(searchText)")
+                            let isFollowing = userEnvironmentData.user.followings.contains { user in
+                                user.name == searchText
+                            }
+                            print(isFollowing)
+                        }
+                    }
+                if isFocused {
+                    Button(action: {
+                        self.searchText = ""
+                    }) {
+                        Image(systemName: "multiply.circle.fill")
+                            .foregroundColor(.gray)
+                    }
+                }
             }
             .padding()
             .background(Color(.systemGray5))
             .cornerRadius(15)
             .padding(.horizontal)
-
+            
             if userEnvironmentData.user.followings.isEmpty {
                 VStack {
                     Text("フレンドがいません。\"**アカウント < フレンド検索**\"からフレンドを追加してください。")
                 }
                 .padding()
             }
-
+            
             List(filteredFollowings, id: \.self) { user in
                 let cardColors = Colors.cardColors
                 UserCard(userName: user.name, color: cardColors[userEnvironmentData.user.followings.firstIndex(of: user)! % cardColors.count], action: {
@@ -77,15 +79,28 @@ struct HomeView: View {
                 })
                 .swipeActions(edge: .trailing) {
                     Button("削除", role: .destructive) {
-                        viewModel2.DeleteFriend(friendId: user._id)
-                        userEnvironmentData.user.followings.removeAll { $0._id == user._id }
-                            }
-                        }
+                        userToDelete = user
+                        isAlertPresented = true
+                    }
+                }
                 .listRowInsets(EdgeInsets())
             }
             .listStyle(GroupedListStyle())
             .scrollContentBackground(.hidden)
             .background(Color.white)
+            .alert(isPresented: $isAlertPresented){
+                Alert(
+                    title: Text("フレンドを削除しますか？"),
+                    primaryButton: .cancel(Text("キャンセル")),
+                    secondaryButton: .destructive(Text("削除"), action: {
+                        if let user = userToDelete {
+                            print("削除完了")
+                            viewModel2.DeleteFriend(friendId: user._id)
+                            userEnvironmentData.user.followings.removeAll { $0._id == user._id }
+                        }
+                    })
+                )
+            }
         }
     }
 }
